Moved CollectionId out of api
