+++ src/api/message.rs
@@ -2,9 +2,8 @@ use common::*;
 
 pub trait Message: Sized {
   type Protobuf;
-  type Error: Debug;
 
-  fn from_protobuf(protobuf: Self::Protobuf) -> Result<Self, Self::Error>;
+  fn from_protobuf(protobuf: Self::Protobuf) -> Result<Self, api::Error>;
 
   fn into_protobuf(self) -> Self::Protobuf;
